refactor: rename mdx components file
